1. Clean up all warnings for C source files in Base tool project. 2. Turn on warning as error for GCC and MSFT CFLAGS so that build will break if a developer introduces a warning during development. 3. Increase waring level to /W4 for MSFT tool chain to detect more warnings in C source code.
git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@1665 c2973edb-eda0-4c78-bc6a-9341b269661f

+++ Source/C/Common/EfiUtilityMsgs.c
@@ -34,7 +34,7 @@ Abstract:
 //
 STATIC STATUS mStatus                 = STATUS_SUCCESS;
 STATIC CHAR8  mUtilityName[50]        = { 0 };
-STATIC UINT32 mPrintLogLevel          = INFO_LOG_LEVEL;
+STATIC UINT64 mPrintLogLevel          = INFO_LOG_LEVEL;
 STATIC CHAR8  *mSourceFileName        = NULL;
 STATIC UINT32 mSourceFileLineNum      = 0;
 STATIC UINT32 mErrorCount             = 0;
@@ -355,7 +355,7 @@ VOID
 DebugMsg (
   CHAR8   *FileName,
   UINT32  LineNumber,
-  UINT32  MsgLevel,
+  UINT64  MsgLevel,
   CHAR8   *Text,
   CHAR8   *MsgFmt,
   ...
@@ -665,7 +665,7 @@ Returns:
 
 VOID
 SetPrintLevel (
-  UINT32  LogLevel
+  UINT64  LogLevel
   )
 /*++
